fix (doesn't help)

+++ src/glove.c
@@ -36,8 +36,8 @@
 typedef long double real;
 
 typedef struct cooccur_rec {
-    int word1;
-    int word2;
+    long int word1;
+    long int word2;
     real val;
 } CREC;
